+++ paleomix/pipelines/phylo/config.py
@@ -52,7 +52,7 @@ def build_parser():
         "--version", action="version", version="%(prog)s v" + paleomix.__version__,
     )
 
-    paleomix.common.logging.add_argument_group(parser, default="info")
+    paleomix.common.logging.add_argument_group(parser)
 
     group = parser.add_argument_group("Scheduling")
     group.add_argument(